requirements.txt
bs4
emails
IPython
jupyterlab
pandas
python-magic